pub_sub/python/sdk/order-processor-fastapi/requirements.txt
fastapi
dapr-ext-fastapi==1.5.0
cloudevents
uvicorn
